Increasing size only on one side

+++ Assets/Scripts/GameManager.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour {
 
-    public int qq = 0;
-    public int ww = 0;
-    public int ee = 0;
-    public int rr = 0;
-    public int vv = 0;
-    public int uu = 0;
-    public int ii = 0;
-    public int oo = 0;
-    public int pp = 0;
-    public int nn = 0;
+    public float qq = 0;
+    public float ww = 0;
+    public float ee = 0;
+    public float rr = 0;
+    public float vv = 0;
+    public float uu = 0;
+    public float ii = 0;
+    public float oo = 0;
+    public float pp = 0;
+    public float nn = 0;
     
     public bool keyPressed = false;
 
@@ -28,61 +28,61 @@ public class GameManager : MonoBehaviour {
         if (Input.GetKeyDown(KeyCode.Q))
         {
             keyPressed = true;
-            qq += 1;
+            qq += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.W))
         {
             keyPressed = true;
-            ww += 1;
+            ww += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.E))
         {
             keyPressed = true;
-            ee += 1;
+            ee += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.R))
         {
             keyPressed = true;
-            rr += 1;
+            rr += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.V))
         {
             keyPressed = true;
-            vv += 1;
+            vv += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.U))
         {
             keyPressed = true;
-            uu += 1;
+            uu += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.I))
         {
             keyPressed = true;
-            ii += 1;
+            ii += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.O))
         {
             keyPressed = true;
-            oo += 1;
+            oo += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.P))
         {
             keyPressed = true;
-            pp += 1;
+            pp += 1f;
         }
 
         if (Input.GetKeyDown(KeyCode.N))
         {
             keyPressed = true;
-            nn += 1;
+            nn += 1f;
         }
 
         keyPressed = false;
@@ -93,44 +93,44 @@ public class GameManager : MonoBehaviour {
     {
         if (qq > 0 && !keyPressed)
         {
-            qq -= 1;
+            qq -= 1f;
         }
         if (ww > 0 && !keyPressed)
         {
-            ww -= 1;
+            ww -= 1f;
         }
         if (ee > 0 && !keyPressed)
         {
-            ee -= 1;
+            ee -= 1f;
         }
         if (rr > 0 && !keyPressed)
         {
-            rr -= 1;
+            rr -= 1f;
         }
         if (vv > 0 && !keyPressed)
         {
-            vv -= 1;
+            vv -= 1f;
         }
 
         if (uu > 0 && !keyPressed)
         {
-            uu -= 1;
+            uu -= 1f;
         }
         if (ii > 0 && !keyPressed)
         {
-            ii -= 1;
+            ii -= 1f;
         }
         if (oo > 0 && !keyPressed)
         {
-            oo -= 1;
+            oo -= 1f;
         }
         if (pp > 0 && !keyPressed)
         {
-            pp -= 1;
+            pp -= 1f;
         }
         if (nn > 0 && !keyPressed)
         {
-            nn -= 1;
+            nn -= 1f;
         }
     }
 }
